[VAS-1095] create jwt
