--- e2e/tsconfig.e2e.json
@@ -1,6 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es6",
-    "module": "commonjs"
-  }
-}
